Support both new and old port format
Arukas API will change their port format which is the same as Docker
Compose settings.
It will support the old format for a while for requesting - creating
an app with services - but it will always respond with new format.
So supporting both new and old port format, you don't need to be
nervous when to change code.

After Arukas API released the new format, the old format converting
part will be redudant and will be fine to removed.